refactor: local cart service works

+++ src/app/features/checkout/components/checkout-product/checkout-product.component.ts
@@ -5,6 +5,9 @@ import { Router, RouterLink } from '@angular/router';
 import { NonNullableFormBuilder, Validators } from '@angular/forms';
 import { DecimalPipe } from '@angular/common';
 import { rxResource } from '@angular/core/rxjs-interop';
+import { AuthService } from '../../../auth/services/auth.service';
+import { LocalCartService } from '../../../home/services/local-cart.service';
+import { of } from 'rxjs';
 
 @Component({
   selector: 'app-checkout-product',
@@ -16,44 +19,75 @@ export class CheckoutProductComponent {
   private readonly apiCartService = inject(ApiCartService);
   private readonly fb = inject(NonNullableFormBuilder);
   private readonly router = inject(Router);
+  private readonly authService = inject(AuthService);
+  private readonly localCartService = inject(LocalCartService);
+
+  readonly auth = this.authService.isAuthenticated();
 
   readonly cartCount = this.apiCartService.cartCount;
 
   readonly cartResource = rxResource({
-    stream: () => this.apiCartService.getCart(),
+    stream: () => {
+      if (this.auth) {
+        return this.apiCartService.getCart();
+      }
+      return of(this.localCartService.getCart())
+    }
   });
 
   increaseQuantity(productId: number) {
-    this.apiCartService.addItem(productId).subscribe({
-    next: (cart) => {
-      this.cartResource.set(cart);
-    },
-    error: (error) => {
-      console.error('Error al agregar producto +:', error);
+    if (this.auth) {
+      //user authenticated
+      this.apiCartService.addItem(productId).subscribe({
+        next: () => {
+          this.cartResource.reload();
+        },
+        error: (error) => {
+          console.error('Error al agregar producto +:', error);
+        }
+      });
+    } else {
+      //user no authenticated
+      this.localCartService.increaseQuantity(productId);
+      this.cartResource.reload();
     }
-  });
   }
 
-  decreaseQuantity(productId: number)  {
-    this.apiCartService.decreaseItem(productId).subscribe({
-    next: (cart) => {
-      this.cartResource.set(cart);
-    },
-    error: (error) => {
-      console.error('Error al disminuir -:', error);
+  decreaseQuantity(productId: number) {
+    if(this.auth){
+      //user authenticated
+      this.apiCartService.decreaseItem(productId).subscribe({
+      next: () => {
+        this.cartResource.reload();
+      },
+      error: (error) => {
+        console.error('Error al disminuir -:', error);
+      }
+    });
+    }else{
+      //user no authenticated
+      this.localCartService.decreaseQuantity(productId);
+      this.cartResource.reload();
     }
-  });
+
   }
 
-  removeItem(productId: number)  {
-    this.apiCartService.removeItem(productId).subscribe({
-    next: (cart) => {
-      this.cartResource.set(cart);
-    },
-    error: (error) => {
-      console.error('Error al remove:', error);
+  removeItem(productId: number) {
+    if(this.auth){
+      //user authenticated
+      this.apiCartService.removeItem(productId).subscribe({
+      next: (cart) => {
+        this.cartResource.set(cart);
+      },
+      error: (error) => {
+        console.error('Error al remove:', error);
+      }
+    });
+    }else{
+      //user no authenticated
+      this.localCartService.removeFromCart(productId);
+      this.cartResource.reload();
     }
-  });
   }
 
   readonly formulario = this.fb.group({
